+++ Tests/IntegrationTestBundle/Controller/FormFlowController.php
@@ -133,7 +133,7 @@ class FormFlowController extends Controller {
 		}
 
 		if ($flow->redirectAfterSubmit($submittedForm)) {
-			$request = $this->getCurrentRequest();
+			$request = $this->get('request_stack')->getCurrentRequest();
 			$params = $this->get('craue_formflow_util')->addRouteParameters(array_merge($request->query->all(),
 					$request->attributes->get('_route_params')), $flow);
 
@@ -145,17 +145,6 @@ class FormFlowController extends Controller {
 			'flow' => $flow,
 			'formData' => $formData,
 		));
-	}
-
-	protected function getCurrentRequest() {
-		if ($this->has('request_stack')) {
-			return $this->get('request_stack')->getCurrentRequest();
-		}
-
-		// TODO remove as soon as Symfony >= 2.4 is required
-		if ($this->has('request')) {
-			return $this->get('request');
-		}
 	}
 
 }