+++ Tests/IntegrationTestBundle/Form/CreateTopicForm.php
@@ -20,7 +20,6 @@ class CreateTopicForm extends AbstractType {
 	 */
 	public function buildForm(FormBuilderInterface $builder, array $options) {
 		$useFqcn = method_exists('Symfony\Component\Form\AbstractType', 'getBlockPrefix'); // Symfony's Form component >=2.8
-		$usePlaceholder = method_exists('Symfony\Component\Form\AbstractType', 'configureOptions'); // Symfony's Form component 2.6 deprecated the "empty_value" option, but there seems to be no way to detect that version, so stick to this >=2.7 check.
 		$isBugReport = $options['isBugReport'];
 
 		switch ($options['flow_step']) {
@@ -36,7 +35,7 @@ class CreateTopicForm extends AbstractType {
 				$choices = Topic::getValidCategories();
 				$builder->add('category', $useFqcn ? 'Symfony\Component\Form\Extension\Core\Type\ChoiceType' : 'choice', array_merge($defaultChoiceOptions, array(
 					'choices' => array_combine($choices, $choices),
-					$usePlaceholder ? 'placeholder' : 'empty_value' => '',
+					'placeholder' => '',
 				)));
 				break;
 			case 2: