+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - [#145]: bumped Symfony dependency to 2.3
   - [#148]: restructured data storage
   - [#180]: renamed step config option `type` to `form_type`
+  - [#184]: made the bundle Symfony 3 compatible, bumped Symfony dependency to 2.6
   - removed the step field template
   - renamed property `step` to `stepNumber` and method `getStep` to `getStepNumber` within event classes
 - [#98]+[#143]: add a validation error to the current form if a form of a previous step became invalid
@@ -42,6 +43,7 @@
 [#175]: https://github.com/craue/CraueFormFlowBundle/issues/175
 [#178]: https://github.com/craue/CraueFormFlowBundle/issues/178
 [#180]: https://github.com/craue/CraueFormFlowBundle/issues/180
+[#184]: https://github.com/craue/CraueFormFlowBundle/issues/184
 [#196]: https://github.com/craue/CraueFormFlowBundle/issues/196
 
 ## 2.1.8 (2015-06-11)